Revert translator name
